Clean up AI generated code (#11576)
fix(core): Manually cleaning up AI code

fix(core): Manually cleaning up AI code

fix(core): Manually cleaning up AI code

fix(core): Manually cleaning up AI code

fix(core): Manually cleaning up AI code

fix(core): Manually cleaning up AI code

fix(core): Manually cleaning up AI code

fix(core): Manually cleaning up AI code

+++ dd-trace-core/src/test/java/datadog/trace/core/propagation/NoneHttpExtractorTest.java
@@ -94,7 +94,7 @@ class NoneHttpExtractorTest extends DDJavaSpecification {
   void extractHttpHeaders(
       String traceId,
       String spanId,
-      @ConvertWith(PrioritySamplingConverter.class) int samplingPriority,
+      @ConvertWith(PrioritySamplingConverter.class) byte samplingPriority,
       String origin,
       boolean allowComma) {
     injectEnvConfig(
